requirement.txt
alembic==1.7.1
APScheduler==3.7.0
blinker==1.4
captcha==0.3
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-APScheduler==1.12.2
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
greenlet==1.1.1
importlib-metadata==4.8.1
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
Pillow==8.3.2
psutil==5.8.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.23
sqlparse==0.4.2
typing-extensions==3.10.0.2
tzlocal==2.1
Werkzeug==2.0.1
zipp==3.5.0
